Allow participant search to be case insensitive.
In order to test this, I needed a better participant test, so I split
the participant_client into a simple example client, and a real test.
Which got me pulling on the thread that lead to me doing the
unit-testing refactor in this commit.

Now all real tests are in the test directory.  With a (very concise) README.