[DA-161] Add test for PhysicalMeasurementsAPI exercising handler and datastore in memory. (#196)
*   Remove PUT and PATCH handlers from PhysicalMeasurementsAPI.
*   Move data-loading test functions to shared location.
*   Update testing docs, misc style cleanup.
This is preparation for updating PhysicalMeasurementsAPI behavior (amending old measurements [DA-161], dropping excess validation [DA-124]). This also proves out a new test setup (using an in-memory flask server instead of a separate server as used by client tests) which I plan to make a test base class for in the future.