[DA-300,DA-283] Consent PDF validation: allow leading slash or not on PDF path in extension. Include BadRequestion.description in error log. (#402)
The error log was masking the exception details, so this should get us better debugging info.

The current code was requiring a leading slash on the PDF's path in the extension, whereas the request didn't have that. I think that was the error. I changed the code to accept either format (leading slash or no).